refactor(hush-cli): extract init and scaffold from pkg_cli
Move cmd_pkg_init into pkg_cli/init.rs and the template writers +
content generators (write_template_file, scaffold_*, generate_*,
sanitize_cargo_package_name, default_guard_wasm_entrypoint) into
pkg_cli/scaffold.rs. cmd_pkg_init and scaffold_package widened to
pub(super); sanitize_cargo_package_name widened to pub(super) for the
wasm-gated find_wasm_binary helper (mod.rs import carries the matching
#[cfg(feature = "wasm-plugin-runtime")]). PLUGIN_MANIFEST_FILENAME
widened to pub(super) const for scaffold's use.
